tests/requirements.txt
aiohttp==3.8.4
pytest==7.2.1
pytest_httpserver==1.0.6
python-dotenv==1.0.0
requests==2.28.0
supergood==1.0.6
supergood.egg==info
urllib3==1.26.9
Werkzeug==2.2.3
